refactor(eda.create_db_report): clean up code
