+++ dataprep/datasets/__init__.py
@@ -1,6 +1,6 @@
 """This module implements load dataset related functions"""
 
-from ._base import load_dataset, _load_dataset_as_dask, load_sqlite_db_url
+from ._base import load_dataset, _load_dataset_as_dask, load_db
 from ._base import get_dataset_names
 
-__all__ = ["load_dataset", "get_dataset_names", "_load_dataset_as_dask", "load_sqlite_db_url"]
+__all__ = ["load_dataset", "get_dataset_names", "_load_dataset_as_dask", "load_db"]
