+++ dataprep/eda/create_db_report/model/init_database.py
@@ -1,7 +1,3 @@
-from .init_table import Table
-from .db_metadata import DbMeta
-
-
 class Database:
     def __init__(self, databaseName: str, schema: str, database_stats) -> None:
         self.databaseName = databaseName
