+++ dataprep/eda/create_db_report/views/pagedata.py
@@ -1,6 +1,3 @@
-from tokenize import String
-
-
 class PageData:
     def __init__(self, templateName: str, scriptName: str):
         self.templateName = templateName
