+++ dataprep/eda/create_db_report/views/pystache_columns.py
@@ -13,7 +13,7 @@ class MustacheTableColumn:
             keyType = " class='primaryKey' title='Primary Key'"
         elif self.tableColumn.isForeignKey():
             keyType = " class='foreignKey' title='Foreign Key'"
-        elif self.indexColumn():
+        elif self.indexColumn:
             keyType = " class='" + self.markAsIndexColumn() + "' title='Indexed'"
 
         return keyType
@@ -24,7 +24,7 @@ class MustacheTableColumn:
             keyTitle = "Primary Key"
         elif self.tableColumn.isForeignKey():
             keyTitle = "Foreign Key"
-        elif self.indexColumn():
+        elif self.indexColumn:
             keyTitle = "Indexed"
 
         return keyTitle
@@ -35,7 +35,7 @@ class MustacheTableColumn:
             keyClass = "primaryKey"
         elif self.tableColumn.isForeignKey():
             keyClass = "foreignKey"
-        elif self.indexColumn():
+        elif self.indexColumn:
             keyClass = "indexedColumn"
 
         return keyClass
@@ -44,7 +44,7 @@ class MustacheTableColumn:
         keyIcon = ""
         if self.tableColumn.isPrimary() or self.tableColumn.isForeignKey():
             keyIcon = "<i class='icon ion-key iconkey' style='padding-left: 5px;'></i>"
-        elif self.indexColumn():
+        elif self.indexColumn:
             keyIcon = "<i class='fa fa-sitemap fa-rotate-120' style='padding-right: 5px;'></i>"
 
         return keyIcon
